[mlir] Fix warnings
This patch fixes:

  mlir/lib/Dialect/Linalg/IR/LinalgOps.cpp:1348:31: error: comparison
  of integers of different signs: 'size_t' (aka 'unsigned long') and
  'int64_t' (aka 'long') [-Werror,-Wsign-compare]

  mlir/lib/ExecutionEngine/SparseTensor/File.cpp:110:3: error: default
  label in switch which covers all enumeration values
  [-Werror,-Wcovered-switch-default]

+++ mlir/lib/Dialect/Linalg/IR/LinalgOps.cpp
@@ -1345,7 +1345,7 @@ LogicalResult ReduceOp::verify() {
       reducedInputDims.push_back(en.value());
   }
 
-  if (reducedInputDims.size() != initType.getRank()) {
+  if (reducedInputDims.size() != static_cast<size_t>(initType.getRank())) {
     return emitOpError() << "number of dimensions after reduction "
                          << reducedInputDims.size()
                          << " doesn't match the init rank "
